Jews and Mongols
In the current version of PB, there would be bugs if there are Jewish rulers around when the Mongol doomstacks appear (I forgot exactly what but either it'll give them a gazillion troops or give the hordes a gazillion troops if the player is Jewish or something), this was a bug in vanilla 2.0 that was fixed in 2.01. I thus fixed the relevant events.

Also I made sure the Mongols sack Rome event now has the right event picture as in vanilla.