Loans: Fixes, annual payment event revamp, tracing support
- Fixed the potential clause in repay_medium_loan & repay_large_loan
preventing the new pay-by-halves functionality from working in those
cases.
- Updated ai_will_do war checks to include whether any liege is at war
and has raised the character's levies (in addition to a war in which the
character is actually an attacker or defender / war = yes) for all loan
decisions.
- Outfitted annual interest payment event with significant amount of
ai_chance modifiers to make a relatively intelligent decision about
whether to only make the yearly minimum payment(s) or to try to pay-back
as much as possible.
- Instrumented all instances of loan payment (interest), loan taken
(also as a result of the pay-by-halves strategy), and loan paid (or
expelled) events, categorized by loan tier, to allow optional tracing of
AI debt management behavior in the aggregate through, e.g., my batch
savegame series analyzer.