Pope, holy order, merc wins crusade: always gift to new character
This is to prevent general weirdness with inheritance of unrelated
lands.  New character will still be of same culture and religion of
giver, as usual, so there will still be an Italian (or Frankish or
Bedouin, etc.)  ruler put in place in these cases.  Non-AI holy order
or mercenary characters will still get the usual opportunities to give
the crusader state away, as weird (but possible) as that case is.