No more duplicate depose_antipope CBs
You will note in the ChangeLog that I've essentially forced Meneth to
make a new commit whenever he prepares a release, because he will have
to modify the ChangeLog to not only remove [BETA-ONLY ISSUE] lines, if
present, but also remove the [BETA] I've slapped next to the current
beta version label.

In light of a couple instances of really good beta feedback, it seemed
important to at least temporarily highlight current beta issues solved
with credit given to those that nailed the problem for us. People are
catching the GitHub/beta bug and frequently updating their copy, so it
seems worth the extra commit. Besides, it might be worthwhile for us to
give a "Many thanks for feedback during the beta to X, Y, and Z," to
our release announcement when relevant.

Will revert ChangeLog if desired.