Heavy fixes to IR war checks / follows CB
Assumes the SWMH Imperial Reconquest CB is Pure Truth. I couldn't
reasonably question it when aligning the decision logic and testing,
but I can say that all of the de jure liege title switcheroos which we
do by event from 1066-1100 really complicate matters in terms of
displaying what needs to be displayed to the user usefully.