Some fixes/changes to the AF, mostly for 2.1
No vassal king AFs allowed until a workaround is found (one is known
but is ambitious, though tentatively planned as experimental code for
the future).

Vetting with desires_autonomy, which was always a bit broken (e.g.,
it did not get reevaluated), is now over.  The event remains for
historic purposes only and is always = no.  Membership and faction
creation chance no longer depend upon the flag.  Effectively, all
that aren't excluded by the few modifiers in objectives/PB_factions.txt
and that qualify as the right sort of vassals will join now-- and
fast.

Some 2.1 sanity checks to limit possibilities of a meeting chain
starting while the liege has somehow become a leader of some sort
of faction revolt (even the CB overthrow_ruler creates a temp.
title and vassalizes a shit-ton of people now). They shouldn't be
required, but we have something wrong in the codebase, and process
of elimination is possibly pointing the finger this way.  Better to
just delay the meeting until game state makes sense again, insofar
as the current code comprehends it.