Disabled vanilla RoI "convert to province culture" decision for AI
There are already events which do this, and they're smarter about it.
This was the cause of the many complaints about kings and such randomly
changing to obscure cultures due to 1 stray province.  However, note
that the RoI AI decisions to convert to hinduism, buddhism, and jainism
all remain unfettered.