Free vassalized holy orders of wrong religion upon crusade success
This prevents, e.g., the Knights Templar and Knights Hospitaler from
becoming vassals of whichever Caliph wins a Jihad for Jerusalem, if
they were vassal to that title.  That is, their holdings, if any, in
the KoJ would be usurped anyway, but their landless title and the
[in]ability to hire them as holy orders would not be stuck under the
Jihad winner forever, as it was previously.